brute force solution for day 8 part 2
Calculating the least common multiple of the number of steps each start
node takes to reach an end node is sufficient for getting the answer.
I'm not happy with this approach as it makes a number of assumptions
about the input (that each start node will only reach one end node
before repeating; that each start node will always loop back around to
itself), but it feels like this was possibly the kind of solution the
setter had in mind? Perhaps there's an alternative method that can be
achieved by flattening the graph somehow.